feat : 금융 관련 정확한 계산을 위하여 BigDecimal 도입

+++ src/main/java/com/prgms/VoucherApp/domain/Voucher.java
@@ -2,10 +2,11 @@ package com.prgms.VoucherApp.domain;
 
 import com.prgms.VoucherApp.dto.VoucherDto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface Voucher {
-    long discount(long beforeAmount);
+    BigDecimal discount(BigDecimal beforeAmount);
 
     UUID getUUID();
